+++ handler.go
@@ -21,12 +21,25 @@ package logit
 import (
 	"errors"
 	"io"
-	"os"
-	"strconv"
 	"sync"
-	"time"
+)
 
-	"github.com/FishGoddess/logit/wrapper"
+const (
+	// DefaultTimeFormat is the default format for formatting time.
+	DefaultTimeFormat = "2006-01-02 15:04:05"
+)
+
+var (
+	// handlers stores all registered handlers.
+	// mutexOfHandlers is for concurrency.
+	handlers        = map[string]func(params map[string]interface{}) Handler{}
+	mutexOfHandlers = &sync.RWMutex{}
+
+	// HandlerIsExistedError is an error happens on repeating handler name.
+	HandlerIsExistedError = errors.New("the name of handler you want to register already exists! May be you should give it an another name")
+
+	// HandlerIsNotExistedError is an error happens on failing to find the handler.
+	HandlerIsNotExistedError = errors.New("the handler you pointed is not existed! Please check the names of all handlers")
 )
 
 // Handler is an interface representation of log handler.
@@ -42,127 +55,6 @@ type Handler interface {
 	// Then all handlers after current handler will not be used.
 	Handle(log *Log) bool
 }
-
-// WriterOf returns a writer implement with given params.
-// Different writer implement may have different params, so what params should
-// be used is dependent to specific writer implement.
-// An example of this method in config:
-//
-//     "writer": {
-//         "rolling": "duration",
-//         "duration": 1,
-//         "directory": "D:/"
-//     }
-//
-func WriterOf(params map[string]interface{}) io.Writer {
-
-	// 默认使用 os.Stdout
-	writer := os.Stdout
-	param, ok := params["writer"]
-	if !ok {
-		return writer
-	}
-
-	// 下面这段代码有些 “肮脏”，但是大张旗鼓地重构这个又有点主次不分的感觉，所以先保持这样，后续再考虑这个点
-	writerConfig := param.(map[string]interface{})
-	rolling, ok := writerConfig["rolling"]
-	if !ok {
-		return writer
-	}
-
-	switch rolling {
-
-	// 以时间间隔进行滚动的日志写出器
-	case "duration":
-
-		// 滚动的时间间隔，单位是秒
-		duration := 24 * 60 * 60 // 一天
-		if param, ok := writerConfig["duration"]; ok {
-			duration = int(param.(float64))
-		}
-
-		// 写出的目标文件夹
-		directory := "./"
-		if param, ok := writerConfig["directory"]; ok {
-			directory = param.(string)
-		}
-
-		return wrapper.NewDurationRollingFile(time.Duration(duration)*time.Second, wrapper.NextFilename(directory))
-
-	// 以文件大小进行滚动的日志写出器
-	case "size":
-
-		// 滚动的文件大小，单位是 MB
-		size := 64 // 64MB
-		if param, ok := writerConfig["size"]; ok {
-			size = int(param.(float64))
-		}
-
-		// 写出的目标文件夹
-		directory := "./"
-		if param, ok := writerConfig["directory"]; ok {
-			directory = param.(string)
-		}
-
-		return wrapper.NewSizeRollingFile(int64(size)*wrapper.MB, wrapper.NextFilename(directory))
-
-	// 不滚动的日志写出器
-	case "off":
-
-		// 写出的目标文件
-		if param, ok := writerConfig["file"]; ok {
-			file, err := wrapper.NewFile(param.(string))
-			if err != nil {
-				panic(err)
-			}
-			return file
-		}
-
-		file, err := wrapper.NewFile("./logit-" + strconv.FormatInt(time.Now().Unix(), 10) + wrapper.SuffixOfLogFile)
-		if err != nil {
-			panic(err)
-		}
-		return file
-	}
-
-	return writer
-}
-
-func init() {
-
-	// 注册默认日志处理器
-	RegisterHandler("default", func(params map[string]interface{}) Handler {
-		timeFormat := DefaultTimeFormat
-		if format, ok := params["timeFormat"]; ok && format != "" {
-			timeFormat = format.(string)
-		}
-		return NewDefaultHandler(WriterOf(params), timeFormat)
-	})
-
-	// 注册 Json 格式日志处理器
-	RegisterHandler("json", func(params map[string]interface{}) Handler {
-		timeFormat := ""
-		if format, ok := params["timeFormat"]; ok {
-			timeFormat = format.(string)
-		}
-		return NewJsonHandler(WriterOf(params), timeFormat)
-	})
-}
-
-const (
-	// DefaultTimeFormat is the default format for formatting time.
-	DefaultTimeFormat = "2006-01-02 15:04:05"
-)
-
-var (
-	// handlers stores all registered handlers.
-	// mutexOfHandlers is for concurrency.
-	handlers        = map[string]func(params map[string]interface{}) Handler{}
-	mutexOfHandlers = &sync.RWMutex{}
-
-	// HandlerIsExistedError is an error happens on repeating handler name.
-	HandlerIsExistedError = errors.New("the name of handler you want to register already exists! May be you should give it an another name")
-)
 
 // RegisterHandler registers your handler to logit so that you can use them easily.
 // Return an error if the name is existed, and you should change another name for your handler.
@@ -182,97 +74,42 @@ func RegisterHandler(name string, newHandler func(params map[string]interface{})
 // HandlerOf returns handler whose name is given name and params.
 // Different handler may have different params, so what params should
 // be injected into newHandler is dependent to specific handler.
-// Notice that we don't use an error mechanism or ok mechanism to check the name but
-// a default handler returning mechanism. This is a more convenient way to use handlers (we think).
+// Notice that we use panic mechanism to check the name.
+// This is a more convenient way to use handlers (we think).
 func HandlerOf(name string, params map[string]interface{}) Handler {
 	mutexOfHandlers.RLock()
 	defer mutexOfHandlers.RUnlock()
 	newHandler, ok := handlers[name]
 	if !ok {
-		return NewDefaultHandler(os.Stdout, DefaultTimeFormat)
+		panic(HandlerIsNotExistedError)
 	}
 	return newHandler(params)
 }
 
-// DefaultHandler is a default handler for use.
+// ================================= StandardHandler =================================
+
+// StandardHandler is a standard handler for use.
 // Generally speaking, encoding a log to bytes then written by writer is the most of
-// handlers do. So we provide a default handler, which only need a writer and an encoder.
-//
-// For config:
-//     If you want to use this handler in your logger by config file, try this:
-//
-//         "handlers":{
-//             "default":{
-//
-//             }
-//         }
-//
-//     This will use logit.DefaultTimeFormat to format time, and if you want to
-//     use your layout to format time, try this:
-//
-//         "handlers":{
-//             "default":{
-//                 "timeFormat": "2006-01-02"
-//             }
-//         }
-//
-type DefaultHandler struct {
+// handlers do. So we provide a standard handler, which only need a writer and an encoder.
+// Notice that this handler is not for config file but use in code, so we don't register it.
+type StandardHandler struct {
 	writer     io.Writer
+	encoder    Encoder
 	timeFormat string
 }
 
-// NewDefaultHandler returns a DefaultHandler holder with given writer.
-func NewDefaultHandler(writer io.Writer, timeFormat string) Handler {
-	return &DefaultHandler{
+// NewStandardHandler returns a StandardHandler holder with given writer and encoder.
+func NewStandardHandler(writer io.Writer, encoder Encoder, timeFormat string) Handler {
+	return &StandardHandler{
 		writer:     writer,
+		encoder:    encoder,
 		timeFormat: timeFormat,
 	}
 }
 
 // Handle will encode log and write log by internal writer.
 // Return true so that handlers after it will be used.
-func (dh *DefaultHandler) Handle(log *Log) bool {
-	dh.writer.Write(EncodeToText(log, dh.timeFormat))
-	return true
-}
-
-// JsonHandler is a json handler for use.
-//
-// For config:
-//     If you want to use this handler in your logger by config file, try this:
-//
-//         "handlers":{
-//             "json":{
-//
-//             }
-//         }
-//
-//     This config will not format time, and keep time in unix form. If you want to
-//     use your layout to format time, try this:
-//
-//         "handlers":{
-//             "json":{
-//                 "timeFormat": "2006-01-02"
-//             }
-//         }
-//
-type JsonHandler struct {
-	writer     io.Writer
-	timeFormat string
-}
-
-// NewJsonHandler returns a JsonHandler holder with given writer.
-// If timeFormat == "", then it will not format time and keep time in unix form.
-func NewJsonHandler(writer io.Writer, timeFormat string) Handler {
-	return &JsonHandler{
-		writer:     writer,
-		timeFormat: timeFormat,
-	}
-}
-
-// Handle will encode log and write log by internal writer.
-// Return true so that handlers after it will be used.
-func (jh *JsonHandler) Handle(log *Log) bool {
-	jh.writer.Write(EncodeToJson(log, jh.timeFormat))
+func (sh *StandardHandler) Handle(log *Log) bool {
+	sh.writer.Write(sh.encoder.Encode(log, sh.timeFormat))
 	return true
 }